gtfs_traveltimes

Travel times from a nominated station departing at a nominated time to every other reachable station in a system.

gtfs_traveltimes(
  gtfs,
  from,
  start_time,
  day = NULL,
  from_is_id = FALSE,
  grep_fixed = TRUE,
  route_pattern = NULL,
  minimise_transfers = FALSE,
  prop_stops = 0.5,
  quiet = FALSE
)

Arguments

gtfs
A set of GTFS data returned from extract_gtfs or, for more efficient queries, pre-processed with gtfs_timetable.

from
Name, ID, or approximate (lon, lat) coordinates of start station (as stop_name or stop_id entry in the stops table, or a vector of two numeric values).

start_time
Desired departure time at from station, either in seconds after midnight, a vector of two or three integers (hours, minutes) or (hours, minutes, seconds), an object of class difftime, hms, or lubridate.

day
Day of the week on which to calculate route, either as an unambiguous string (so "tu" and "th" for Tuesday and Thursday), or a number between 1 = Sunday and 7 = Saturday. If not given, the current day will be used. (Not used if gtfs has already been prepared with gtfs_timetable.)

from_is_id
Set to TRUE to enable from parameter to specify entry in stop_id rather than stop_name column of the stops table (same as from_to_are_ids parameter of gtfs_route).

grep_fixed
If FALSE, match station names (when passed as character string) with grep(..., fixed = FALSE), to allow use of grep expressions. This is useful to refine matches in cases where desired stations may match multiple entries.

route_pattern
Using only those routes matching given pattern, for example, "^U" for routes starting with "U" (as commonly used for underground or subway routes. To negate the route_pattern -- that is, to include all routes except those matching the pattern -- prepend the value with "!"; for example "!^U" will include all services except those starting with "U". (This parameter is not used at all if gtfs has already been prepared with gtfs_timetable.)

minimise_transfers
If TRUE, isochrones are calculated with minimal-transfer connections to each end point, even if those connections are slower than alternative connections with transfers.

prop_stops
Stop scanning after this proportion of all potentially reachable stops has been reached. Some stops may only be reached very infrequently (like once per day), and scanning a timetable until all stops have been reached may (1) yield anomalously long travel times for very infrequently serviced stops; and (2) take a long time to calculate. The prop_stops parameter should accordingly generally be less than 1. For large systems with many stops (tens of thousands), values of 0.5 are often sufficient to reach most of the system.

quiet
Set to TRUE to suppress screen messages (currently just regarding timetable construction).

Note

Searching for all connections over an entire timetable may return anomalously high travel times for stops which are only very occasionally serviced. Results generated with values of prop_stops close to 1 may need to be manually cleaned prior to analysis.