summary.gtfs

summary.gtfs

# S3 method for gtfs
summary(object, ...)

Arguments

object
A gtfs object to be summarised

...
ignored here

Examples

berlin_gtfs_to_zip ()
f <- file.path (tempdir (), "vbb.zip")
g <- extract_gtfs (f)

#> ▶ Unzipping GTFS archive

#> ✔ Unzipped GTFS archive  

#> ▶ Extracting GTFS feed

#> ✔ Extracted GTFS feed 

#> ▶ Converting stop times to seconds

#> ✔ Converted stop times to seconds 

#> ▶ Converting transfer times to seconds

#> ✔ Converted transfer times to seconds 

summary (g)

#> A gtfs object with the following tables and respective numbers of entries in each:

#>   calendar     routes      trips stop_times      stops  transfers 
#>       2395         52      28590      22666        957       9833 

g <- gtfs_timetable (g)

#> Day not specified; extracting timetable for wednesday

summary (g) # also summarizes additional timetable information

#> A gtfs timetable object with the following tables and respective numbers of entries in each:

#>   calendar     routes      trips stop_times      stops  transfers  timetable 
#>       2395         52       7086       7626        957       9833       7052 
#>   stop_ids   trip_ids 
#>        957       7086